Use precise32 instead of precise64
... For machines that don't have VT-x support.

But the thing is that the precise32 base box seem to be a little glitchy
and running "service apache2 reload" from Fabric makes apache shut down.

I had to change it for a reload, but even then, sometimes, reload is
glitchy, so I had to add a note to try a second 'fab restart' when it
doesn't work.